requirements.txt
appdirs==1.4.4
asgiref==3.3.0
attrs==20.2.0
cached-property==1.5.2
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-online-users==0.3
django-render-partial==0.4
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
lxml==4.6.1
Pillow==8.0.1
pytz==2020.4
requests==2.24.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
zeep==4.0.0
